ai-provement-tool/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv
groq
Pillow
python-multipart
pytest
httpx